refactoring and adding documentation
